Fixes issue #16
Problem was the auto_clean function for the cache.
When using threads, auto_clean is disabled with set_cache_auto_clean(False) and the cache is manually cleaned with clear_cache_from_expired_items() when all threads are finished.
When blocks() finishes, set_cache_auto_clean(auto_clean) is set to the old value.